Generate the support document for exam correction

From an exam object, we can generate instances that serve as support for the correction of the exam. Each instance will include the answers, if they are indicated, associated with the exercises. In any case, the randomly included part of the exam will be highlighted.

Usage

generate_correction_document(ex, out_dir, output_format, encoding, new_pages)

# S3 method for exam
generate_correction_document(
  ex,
  out_dir = NULL,
  output_format = "pdf_document",
  encoding = "UTF-8",
  new_pages = NULL
)

Arguments

ex

An `exam` object.

out_dir

A string indicating the output folder.

output_format

A vector of strings specifying the desired output formats.

encoding

A string specifying the encoding.

new_pages

A string with the values 'all', 'none', or 'NULL'.

Value

An `exam` object.

Details

To do this, we specify the folder where the documents will be generated (using parameter `out_dir`), the output format (using parameter `output_format`), the encoding (using parameter `encoding`), and whether we want each exercise to start on a new page, include exercises until the pages are filled, or preserve the definition of the exercise in this regard (using parameter `new_pages`).

See also

Examples

# \donttest{
rmd <- system.file("extdata/template01.Rmd", package = "rexer")
exercises <- system.file("extdata/exercises.csv", package = "rexer")
ex <- exam(
  rmd = rmd,
  examinees = NULL,
  instances_num = 1,
  random = TRUE,
  reorder_exercises = TRUE
) |>
  define_exercises_from_csv(exercises) |>
  generate_correction_document(out_dir = tempdir())
#> 
#> 
#> processing file: template01.Rmd
#> 1/5                  
#> 2/5 [setup]          
#> 3/5                  
#> 4/5 [unnamed-chunk-1]
#> 5/5                  
#> output file: template01.knit.md
#> /opt/hostedtoolcache/pandoc/3.1.11/x64/pandoc +RTS -K512m -RTS template01.knit.md --to latex --from markdown+autolink_bare_uris+tex_math_single_backslash --output /tmp/RtmpaWrqXL/1.tex --lua-filter /home/runner/work/_temp/Library/rmarkdown/rmarkdown/lua/pagebreak.lua --lua-filter /home/runner/work/_temp/Library/rmarkdown/rmarkdown/lua/latex-div.lua --embed-resources --standalone --highlight-style tango --pdf-engine pdflatex --variable graphics --include-in-header /tmp/RtmpaWrqXL/rmarkdown-str1fd96990f870.html --include-in-header /tmp/RtmpaWrqXL/rmarkdown-str1fd9282100fd.html 
#> Warning: error in running command
#> ! sh: 1: pdflatex: not found
#> Error: LaTeX failed to compile /tmp/RtmpaWrqXL/1.tex. See https://yihui.org/tinytex/r/#debugging for debugging tips. See 1.log for more info.
# }